Version 2.5.0
= Version 2.5.0 – March 18, 2024 =

+   Only flush transients (flush_caches()) if not using an object cache.
+   Reworked get_option(),get_network_option() for faster access to non-reserved options.
+   Added support for `$_SERVER['GEOIP_COUNTRY_CODE']` IP Geolocation.
+   Use `this.form.requestSubmit()` instead of `options_form.submit()`.
+   Fixed switch_to_blog() and loading/saving options on switch.
+   Updated maintenance mode, enable using transient with expiration.
+   Recognize 'Network Enabled' as extension enabled option.
+   New get_page_template() to buffer a template part.
+   Added action `after_flush_caches` when flushing caches.
+   Added option `{classname}_selected_update_channel` and stdOptions_updateChannel() in standard_options trait.
    +   Like {classname}_PLUGIN_UPDATE_CHANNEL.
+   Added constant `{classname}_PLUGIN_UPDATE_CHANNEL` to override plugin update source.
    +   For github hosting, specify branch|release and the tag or id.
    +   `define( 'EACDOOJIGGER_PLUGIN_UPDATE_CHANNEL', 'branch' );`
    +   `define( 'EACDOOJIGGER_PLUGIN_UPDATE_CHANNEL', 'branch/default' );`
+   Updated documentation and examples.
    +   Removed Extras from distribution package, now available on Github:
        +   [Download](https://swregistry.earthasylum.com/software-updates/eacdoojigger-extras.zip)
        +   [Documentation](https://github.com/EarthAsylum/docs.eacDoojigger/wiki/)
+   Streamline standard_options trait, added checkForUpdates (tools).
+   Streamline plugin_update trait including
    +   Support for new software taxonomy Github hosting plugin.
    +   Renamed methods in plugin_update trait.
    +   Delete internal transient when update_plugins transient is deleted.
    +   Changed default transient_time to 1 hour (was 12), letting WP manage update checks.
+   Use wp_clean_update_cache() when clearing caches.
+   Now loads for all (php/non-php) requests.
    +   Extensions are not loaded unless self::ALLOW_NON_PHP is set on construct.
    +   This allows handling of non-php files and redirects passed through WordPress.
    +   Limits loading of unneeded extensions.
+   Pass authentication header on plugin_update_parameters filter in swRegistrationUI.
+   Added environment to plugin updater uri - ?environment=wp_get_environment_type().
+   Fixed dynamic properties (no longer supported in PHP 8.2+) in anonymous class in abstract_core.
+   Fixed error in debugging_extension::capture_deprecated()
+   Now hosted on and updated from github (private repository).
    +   Using [{eac}SoftwareRegistry](https://swregistry.earthasylum.com/) plugin.
    +   With [Software Product Taxonomy](https://swregistry.earthasylum.com/software-taxonomy/) extension for github hosting.

+++ EarthAsylumConsulting/abstract_core.class.php
@@ -1451,7 +1451,7 @@ abstract class abstract_core
 	public function flush_caches(bool $fullFlush=false): void
 	{
 		$message = 'cache';
-		if (method_exists($this,'deleteTransients') && $fullFlush)
+		if (method_exists($this,'deleteTransients') && $fullFlush && !wp_using_ext_object_cache())
 		{
 			$this->deleteTransients($this->is_network_admin());
 			$this->logDebug('deleteTransients',__METHOD__);
